chore: Move OpenAPI-specific code to `schemathesis.specs.openapi._hypothesis`
