+++ src/schemathesis/__init__.py
@@ -1,9 +1,9 @@
 from . import fixups, hooks
-from ._hypothesis import init_default_strategies, register_string_format
 from .cli import register_check
 from .constants import __version__
 from .loaders import from_dict, from_file, from_path, from_pytest_fixture, from_uri, from_wsgi
 from .models import Case
 from .specs import graphql
+from .specs.openapi._hypothesis import init_default_strategies, register_string_format
 
 init_default_strategies()
