+++ src/schemathesis/specs/openapi/_hypothesis.py
@@ -1,0 +1,28 @@
+from base64 import b64encode
+from typing import Tuple
+
+from hypothesis import strategies as st
+from requests.auth import _basic_auth_str
+
+
+def register_string_format(name: str, strategy: st.SearchStrategy) -> None:
+    """Register a new strategy for generating data for specific string "format"."""
+    if not isinstance(name, str):
+        raise TypeError(f"name must be of type {str}, not {type(name)}")
+    if not isinstance(strategy, st.SearchStrategy):
+        raise TypeError(f"strategy must be of type {st.SearchStrategy}, not {type(strategy)}")
+    from hypothesis_jsonschema._from_schema import STRING_FORMATS  # pylint: disable=import-outside-toplevel
+
+    STRING_FORMATS[name] = strategy
+
+
+def init_default_strategies() -> None:
+    """Register all default "format" strategies."""
+    register_string_format("binary", st.binary())
+    register_string_format("byte", st.binary().map(lambda x: b64encode(x).decode()))
+
+    def make_basic_auth_str(item: Tuple[str, str]) -> str:
+        return _basic_auth_str(*item)
+
+    register_string_format("_basic_auth", st.tuples(st.text(), st.text()).map(make_basic_auth_str))  # type: ignore
+    register_string_format("_bearer_auth", st.text().map("Bearer {}".format))
